+++ novawallet/Modules/DApp/DAppBrowser/Tabs/DAppBrowserTabManager/DAppBrowserTabManager.swift
@@ -352,7 +352,7 @@ extension DAppBrowserTabManager {
             mapper: AnyCoreDataMapper(mapper)
         )
 
-        let renderFilesRepository = RuntimeFilesOperationFactory(
+        let renderFilesRepository = WebViewRenderFilesOperationFactory(
             repository: FileRepository(),
             directoryPath: ApplicationConfig.shared.webPageRenderCachePath
         )
